Fix a bug in flow entry parsing
The parser would crash if a coma (',') was found outside a flow context.
This is now fixed.

While here, improve errors handling API. Now, when an exception is
thrown, the parser is insde a {yaml_parser, Parser} tuple. Thus, the
caller doesn't need to rely on the #yaml_parser record to identify the
exception and therefore, the definition of this record remains private.