yamerl_constr: Use YAML 1.1 node modules when the document is YAML 1.1
Before this change, we would always assume a YAML 1.2 context. To
achieve this, we defer the time we select the schema and the
corresponding modules to the time we receive a `#yamerl_doc_start{}`
token. If the schema option is set to `auto` (the default), we use the
document version to decide between YAML 1.1 modules or the YAML 1.2 core
schema modules.

For YAML 1.0, the YAML 1.1 modules are selected. For anything after YAML
1.2, the YAML 1.2 modules are selected.

This change fixes another possible problem: if multiple documents in the
same stream use different YAML versions, they should be parsed
correctly.

References #25.
Fixes #27.