src/yamerl.app.src: Bump version to 0.7.0
Thie version now loads the appropriate modules when a document indicates
it follows YAML 1.1. Before, YAML 1.2 core schema modules would always
be used, except if the caller explicitely set the schema to `yaml11`.

Another consequence of this change is that multiple documents in a
single stream should be parsed according to their YAML version if they
specify a different one.

Last, `yamerl_constr` takes a new option, `ignore_unrecognized_tags`, to
ignore unrecognized tags (!) and construct a term as if the node was a
plain untagged node.