yamerl_parser: A flow collection can start on the same line as a tag
To determine if a tag should be attached to a token, we had the
following conditions:
1.  To be attached to a collection start, the tag must have been a
    previous line.
2.  To be attached to a scalar, the tag must have been on the same line
    or any previous ones.

This didn't work for flow collections: the tag could be on the same line
as the character starting the flow collection. The first condition above
ruled out this case, leading to the tag being attached to the first item
or key in the collection.

Now there is a new condition between conditions 1 and 2 above: to be
attached to a flow collection start, the tag must have been on the same
line or any previous ones.

Fixes #34.