yamerl_parser: Fix detection of a block scalar's end
When a block scalar was ended by a document end marker, it was
considered an error because of the unexpected indentation (the
indentation was less than the body of the block scalar).

The problem was that this error condition was verified before checking
this was a document or directives end marker.

Simply swapping the two code blocks fixes the issue.

Fixes #41.