GitHub Actions: Replace workflows by a single one
[Why]
Before, we had the following workflows:
* one to test yamerl
* one to generage and publish docs
* one to publish a release to Hex.pm

One problem was that the docs and the release workflows didn't depend on
the test results.

Another one was that the docs was only built for the master branch before
the publish. So if a pull request broke the docs, we would not notice it
before the breakage was in master.

[How]
The new workflow was copied from horus where the initial work happened.
All workflows are combined into a single workflow with the correct
dependencies.

Also, the docs are always generated regardless of the branch and
published only it is master.